Moved App.js file to components folder and added BooksList and BooksForm in a div selector
